pyproject.toml
[project]
name = "healthai"
version = "0.1.0"
description = "HealthAI - Intelligent Healthcare Assistant powered by AI"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[tool.coverage.run]
source = ["backend", "validation", "config", "ai_client"]
omit = ["*/tests/*", "*/.venv/*"]
